MlBatchedTimer: refactor batched timer updates to a separate class
MlBatchedTimer is a MainLoop dependendent class that batches the updates
to be applied to a timer. Since timer-updates must make a round-trip to
the main thread, performing it is quite expensive, especially latency wise.

This class skips most of the updates, so that timer-updates can be performed
in worker threads, a skipped update is a lockless operation (at the cost
of some inaccuracy).

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>